dancerr-backend/pyproject.toml
[project]
name = "dancerr-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Thibault Viglino", email = "thibault@thvi.ch"},
]
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "PyYAML>=6.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
